Corrections

+++ print_hex.c
@@ -10,12 +10,12 @@
 int print_hex(int n, char *format, va_list args)
 {
 	int count = 0, i;
-	unsigned int num;
+	unsigned long int num;
 	char *str;
 
 	for (; n > 0; n--)
 		va_arg(args, void *);
-	num = va_arg(args, unsigned int);
+	num = va_arg(args, unsigned long int);
 
 	str = to_hex(num);
 	for (i = 0; format[i] != '\0'; i++)
